+++ src/core/structsos/__init__.py
@@ -1,10 +1,12 @@
 from .structsos import StructuralSOS, _structural_sos
 from .utils import radsimp, quadratic_weighting, zip_longest
+from .pivoting import prove_univariate, prove_univariate_interval
 from .solution import SolutionStructural, SolutionStructuralSimple
 from ..shared import SS
 
 __all__ = [
     'StructuralSOS', 'radsimp', 'quadratic_weighting', 'zip_longest',
+    'prove_univariate', 'prove_univariate_interval',
     'SolutionStructural', 'SolutionStructuralSimple'
 ]
