+++ src/core/structsos/pivoting/__init__.py
@@ -1,0 +1,8 @@
+from .univariate import prove_univariate, prove_univariate_interval
+from .bivariate import structural_sos_2vars
+
+__all__ = [
+    'prove_univariate',
+    'prove_univariate_interval',
+    'structural_sos_2vars'
+]
