+++ src/core/symsos/__init__.py
@@ -1,11 +1,11 @@
 from .symsos import (
     SymmetricSOS, sym_representation, TRANSLATION_POSITIVE, TRANSLATION_REAL
 )
-from .univariate import prove_univariate, prove_univariate_interval, check_univariate
+from .representation import prove_univariate, prove_univariate_interval
 from .solution import SolutionSymmetric, SolutionSymmetricSimple
 
 __all__ = [
     'SymmetricSOS', 'sym_representation', 'TRANSLATION_POSITIVE', 'TRANSLATION_REAL',
-    'prove_univariate', 'prove_univariate_interval', 'check_univariate',
+    'prove_univariate', 'prove_univariate_interval',
     'SolutionSymmetric', 'SolutionSymmetricSimple'
 ]
